requirements.txt
pydantic
openai
faiss-cpu
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
asn1crypto==1.5.1
Authlib==1.3.0
backoff==2.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.8
DateTime==5.5
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
joblib==1.4.2
jwt==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
openai==1.35.13
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pg8000==1.24.2
platformdirs==4.2.2
psycopg2-binary==2.9.9
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
scikit-learn==1.5.1
scipy==1.14.0
scramp==1.4.5
setuptools==70.3.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.11.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.38
starlette==0.37.2
threadpoolctl==3.5.0
tomlkit==0.12.5
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
websockets==12.0
zope.interface==6.4.post2
